share the create graphql client logic
