+++ packages/client/src/main.ce.tsx
@@ -1,32 +1,18 @@
 import * as React from 'react';
 import ReactDOM from 'react-dom';
 import {ApolloProvider} from 'react-apollo';
-import {genClient} from 'canner/lib/components/index';
-import GraphqlClient from 'canner-graphql-interface/lib/graphqlClient/graphqlClient';
-import {LocalStorageConnector} from 'canner-graphql-interface';
 import {notification, Button} from 'antd';
 import {BrowserRouter, Route} from 'react-router-dom';
 import {BackgroundTokenSyncer} from './workers/backgroundTokenSyncer';
-import Main, { MainPageSidebarItem } from 'containers/mainPage';
+import MainPage, { MainPageSidebarItem } from 'containers/mainPage';
 import { appPrefix } from 'utils/env';
+import { createGraphqlClient } from 'utils/graphqlClient';
 
 // Icons
 import iconJupyterHub from 'images/icon-jupyterhub.svg'
 
 // Components
 import Jupyterhub from 'containers/jupyterhubPage';
-
-const graphqlClient = new GraphqlClient({
-  uri: (window as any).graphqlEndpoint,
-  fetch: (uri, options) => {
-    const token = window.localStorage.getItem('canner.accessToken');
-    options.headers = {
-      Authorization: `Bearer ${token}`,
-      ...options.headers || {}
-    };
-    return fetch(uri, options);
-  },
-});
 
 const fakeData = {
   me: {
@@ -93,23 +79,19 @@ const fakeData = {
       }]
     }]
   },
-}
+};
 
-const connector = new LocalStorageConnector({
-  defaultData: fakeData,
-  localStorageKey: 'infuse-hub'
-})
+const schema = {
+  me: {type: 'object'},
+};
 
-const client = genClient(process.env.NODE_ENV === 'production' ?
-  {graphqlClient} :
-  {
-    connector,
-    schema: {
-      me: {type: 'object'},
-    }
-  });
+const client = createGraphqlClient({
+  fakeData,
+  schema
+});
 
-class Hub extends React.Component {
+
+class Main extends React.Component {
   render() {
     const sidebarItems: MainPageSidebarItem[] = [
       {
@@ -122,12 +104,12 @@ class Hub extends React.Component {
     return (
       <BrowserRouter>
         <ApolloProvider client={client}>
-          <Main sidebarItems={sidebarItems}>
+          <MainPage sidebarItems={sidebarItems}>
             {/* Jupyterhub */}
             <Route path={`${appPrefix}g/:groupName/hub`} exact>
               <Jupyterhub />
             </Route>
-          </Main>
+          </MainPage>
         </ApolloProvider>
       </BrowserRouter>
     )
@@ -182,5 +164,5 @@ tokenSyncWorker.run().catch(console.error);
 
 // render
 ReactDOM.render(
-  <Hub />
+  <Main />
 , document.getElementById('root'));
